Fix initial codegen for anonymous projects
